# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agent Orcha API",
        "version": "0.0.3",
        "description": "API for orchestrating agents, workflows, knowledge stores, functions, MCP servers, and LLMs."
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/knowledge/{name}": {
            "get": {
                "summary": "Get Knowledge Store Config",
                "tags": ["Knowledge"],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Knowledge store configuration",
                        "content": {
                            "application/json": {
                                "schema": { "type": "object" }
                            }
                        }
                    },
                    "404": { "description": "Knowledge store not found" }
                }
            }
        },
        "/api/knowledge/{name}/search": {
            "post": {
                "summary": "Search Knowledge Store",
                "tags": ["Knowledge"],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "query": { "type": "string" },
                                    "k": { "type": "integer", "default": 4 }
                                },
                                "required": ["query"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Search results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "content": { "type": "string" },
                                                    "metadata": { "type": "object" },
                                                    "score": { "type": "number" }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": { "description": "Knowledge store not found" },
                    "500": { "description": "Internal server error" }
                }
            }
        },
        "/api/knowledge/{name}/refresh": {
            "post": {
                "summary": "Refresh Knowledge Store",
                "tags": ["Knowledge"],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Refresh result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": { "type": "boolean" }
                                    }
                                }
                            }
                        }
                    },
                    "500": { "description": "Internal server error" }
                }
            }
        },
        "/api/knowledge/{name}/add": {
            "post": {
                "summary": "Add Documents to Knowledge Store",
                "tags": ["Knowledge"],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "documents": {
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "properties": {
                                                "content": { "type": "string" },
                                                "metadata": { "type": "object" }
                                            },
                                            "required": ["content"]
                                        }
                                    }
                                },
                                "required": ["documents"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Addition result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": { "type": "boolean" },
                                        "added": { "type": "integer" }
                                    }
                                }
                            }
                        }
                    },
                    "404": { "description": "Knowledge store not found" },
                    "500": { "description": "Internal server error" }
                }
            }
        },
        "/api/knowledge/{name}/entities": {
            "get": {
                "summary": "Get GraphRAG Entities",
                "description": "Returns entities extracted from a GraphRAG knowledge store.",
                "tags": ["Knowledge"],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of entities",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": { "type": "string" },
                                            "type": { "type": "string" },
                                            "name": { "type": "string" },
                                            "description": { "type": "string" },
                                            "properties": { "type": "object" }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": { "description": "Knowledge store not found" }
                }
            }
        },
        "/api/knowledge/{name}/communities": {
            "get": {
                "summary": "Get GraphRAG Communities",
                "description": "Returns communities detected in a GraphRAG knowledge store.",
                "tags": ["Knowledge"],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of communities",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": { "type": "string" },
                                            "nodeIds": { "type": "array", "items": { "type": "string" } },
                                            "summary": { "type": "string" },
                                            "title": { "type": "string" }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": { "description": "Knowledge store not found" }
                }
            }
        },
        "/api/knowledge/{name}/edges": {
            "get": {
                "summary": "Get GraphRAG Edges",
                "description": "Returns edges (relationships) from a GraphRAG knowledge store.",
                "tags": ["Knowledge"],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of edges",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": { "type": "string" },
                                            "type": { "type": "string" },
                                            "sourceId": { "type": "string" },
                                            "targetId": { "type": "string" },
                                            "description": { "type": "string" },
                                            "weight": { "type": "number" }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": { "description": "Knowledge store not found" }
                }
            }
        },
...
    }
}